Updated Tag Features
Made to retrieve tags and post them in the tag section. Tags are being retrieved, you can see in logcat, but they aren't posting. I believe it is an adapter issue. The add tags is not working, I believe okhttp is giving issues, the url works fine, you can test in a web browser and log shows that the right tag values are present but it won't actually work for some reason